requirements.txt
vosk==0.3.22
numpy
pandas
librosa==0.8.0
moviepy
youtube_dl
youtube_dl
imutils
opencv-python
matplotlib
google-cloud-vision
pydantic
fastapi
uvicorn
pymongo
python-dotenv